Crate libcommon_core

This module is used to export the functions and types of libcommon-minimal. Do not use it directly. Use libcommon instead. The different modules are reexported in libcommon.

Re-exports

pub use libcommon_iop_arcswap as iop_env_arcswap;

Modules

bindings

farch

Unarchive embedded content in C files.

helpers

Utility helpers functions.

log

Rust bindings for the C logging library.

lstr

Type-safe wrappers for C lstr_t with ownership and content type safety.

mem_stack

t_pool implementation and manipulation in Rust.

module

Infrastructure for creating C-compatible modules from Rust.

pstream

Conversion between a C pstream_t and Rust string/bytes types.

sb

Module to interact with sb_t in Rust.

thr

Module to export the thread functions used by libcommon in a safe way

Macros

SB_1k

Create a Sb with a 1KB buffer on the stack.

SB_8k

Create a Sb with a 8KB buffer on the stack.

c_module

Creates a C-compatible module with a typed context.

c_module_method

Declare and define a C module method.

function_name

Macro to get the current function name.

logger_debug

Log a message at debug level.

logger_error

Log a message at error level.

logger_fatal

Log a message at fatal level and exit.

logger_info

Log a message at info level.

logger_log

Log a message at a given level.

logger_notice

Log a message at notice level.

logger_panic

Log a message at panic level and abort.

logger_trace

Log a message at trace level.

logger_warning

Log a message at warning level.

t_SB_1k

Create a Sb with a 1KB buffer on the TScope.

t_SB_8k

Create a Sb with a 8KB buffer on the TScope.