libcommon_core::bindings

Function mem_tool_resize_block

pub unsafe extern "C" fn mem_tool_resize_block(
    mem: *const c_void,
    old_len: usize,
    new_len: usize,
    rz: usize,
)

Indicate the resizing of a memory block previously allocated by a malloc()-like function.

This API allows to handle a basic realloc() behavior. It must be called after the allocation of extra memory (if any) and before the deallocation of old memory (if any).

For Valgrind it has the good property of preserving the V-bits (tracking defined and undefined bytes) of the memory overlap between the old and the new block.

Sadly, there is currently no API that allows to realloc() a block of memory with a change of memory address. In such case you need to freelike() the old address and malloclike() the new address, but you will loose the V-bits of the old memory in the operation.

In case you really need such API, it’s implementable by using VALGRIND_GET_VBITS() on the old memory block and VALGRIND_SET_VBITS() on the new memory block. It’s not easy to map in the mem-tool model which is supposed to handle both Asan and Valgrind. Indeed this is completely specific to Valgrind and would require a set of dedicated API and structures since it cannot be done in a single function. You generally don’t have access to the old memory after a realloc()-like and you don’t have access to the new memory before the realloc()-like never. So one must first store the V-bits of the old memory, then perform the realloc() and then restore the V-bits.

\param[in] mem Memory block address. \param[in] old_len Previous size of the memory block. \param[in] new_len New size of the memory block. \param[in] rz The size of the red zones added around the memory block (see <valgrind.h> for more details). \param[in] defined Whether the memory should be considered defined or undefined. This parameter makes sense only with Valgrind.