libcommon_core::bindings

Function thr_queue

pub unsafe extern "C" fn thr_queue(
    q: *mut thr_queue_t,
    job: *mut thr_job_t,
)

\brief Queue one job on a serial queue

For simplicity, #q can be NULL and then thr_queue* is similar to calling thr_schedule*.

Jobs on serial queues run at most one at a time. A specific queue, called thr_queue_main_g see its jobs run in the main thread only.