libcommon_core

Module mem_stack

t_pool implementation and manipulation in Rust.

Unlike in C the TScope allocator object needs to be passed around to properly associate the lifetimes of the variables.

The destructor is not called for values allocated by the TScope.

Using thr::attach() and thr:detach() are required to use TScope.

Structs

TScope

Rust representation of a TScope