libcommon_core::module

Trait IntoModuleResult

pub trait IntoModuleResult {
    // Required method
    fn into_module_result(self) -> Result<(), Box<dyn Error>>;
}

Trait for converting a callback return value into a module’s result type.

Implemented for () (infallible) and Result<(), E> where E: Into<Box<dyn Error>>. This allows the initialize and shutdown callbacks to return either nothing (when they cannot fail) or any Result<(), E> without forcing the caller to box the error themselves.

Required Methods

fn into_module_result(self) -> Result<(), Box<dyn Error>>

Convert into a Result<(), Box<dyn Error>>.

Errors

Returns the error from the underlying Result, boxed into Box<dyn Error>. The unit (()) implementation never errors.